+++ src/main/kotlin/kr/hhplus/be/server/infrastructure/persistence/statistics/model/entity/PopularProductEntity.kt
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.infrastructure.persistence.statistics.model.entity
 
 import jakarta.persistence.*
-import kr.hhplus.be.server.domain.statistics.product.model.PopularProduct
+import kr.hhplus.be.server.domain.statistics.product.model.entity.PopularProduct
 import kr.hhplus.be.server.infrastructure.persistence.common.entity.BaseEntity
 
 @Entity
